LayoutTests/fast/media/mq-hover-matchMedia.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
    description("Test the hover media feature with machMedia.")

    function testHoverCondition(condition) {
        var testString = 'window.matchMedia("(' + condition + ')").matches';
        debug(testString + " = " + eval(testString));
    }

    testHoverCondition("hover");

    testHoverCondition("hover: hover");
    testHoverCondition("hover: on-demand");
    testHoverCondition("hover: none");

    testHoverCondition("hover: Hover");
    testHoverCondition("hover: On-demand");
    testHoverCondition("hover: None");

    testHoverCondition("Hover: hover");
    testHoverCondition("Hover: on-demand");
    testHoverCondition("Hover: none");
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
