LayoutTests/platform/ios/fast/media/mq-hover-matchMedia-expected.txt

Test the hover media feature with machMedia.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


window.matchMedia("(hover)").matches = false
window.matchMedia("(hover: hover)").matches = false
window.matchMedia("(hover: on-demand)").matches = false
window.matchMedia("(hover: none)").matches = true
window.matchMedia("(hover: Hover)").matches = false
window.matchMedia("(hover: On-demand)").matches = false
window.matchMedia("(hover: None)").matches = true
window.matchMedia("(Hover: hover)").matches = false
window.matchMedia("(Hover: on-demand)").matches = false
window.matchMedia("(Hover: none)").matches = true
PASS successfullyParsed is true

TEST COMPLETE
